LayoutTests/http/tests/dom/noreferrer-window-not-targetable-expected.txt

Make sure that windows opened with 'noreferrer' are not targetable. If testing manually, you should see 2 tabs open.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w.location.href is "about:blank"
PASS testRunner.windowCount() is 2
PASS successfullyParsed is true

TEST COMPLETE
